+++ src/main/java/es/uji/ei102720mgph/SANA/controller/TimeSlotController.java
@@ -2,6 +2,7 @@ package es.uji.ei102720mgph.SANA.controller;
 
 import es.uji.ei102720mgph.SANA.dao.TimeSlotDao;
 import es.uji.ei102720mgph.SANA.model.TimeSlot;
+import es.uji.ei102720mgph.SANA.model.UserLogin;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -10,6 +11,9 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.method.HandlerTypePredicate;
+
+import javax.servlet.http.HttpSession;
 
 @Controller
 @RequestMapping("/timeSlot")
@@ -23,7 +27,12 @@ public class TimeSlotController {
     }
 
     @RequestMapping(value="/add/{naturalArea}")
-    public String addTimeSlot(Model model, @PathVariable String naturalArea) {
+    public String addTimeSlot(Model model, @PathVariable String naturalArea, HttpSession session) {
+        if(session.getAttribute("municipalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/timeSlot/add/" + naturalArea);
+            return "/inicio/login";
+        }
         TimeSlot timeSlot = new TimeSlot();
         timeSlot.setNaturalArea(naturalArea);
         model.addAttribute("timeSlot", timeSlot);
@@ -44,7 +53,12 @@ public class TimeSlotController {
     }
 
     @RequestMapping(value="/update/{id}", method = RequestMethod.GET)
-    public String editTimeSlot(Model model, @PathVariable String id) {
+    public String editTimeSlot(Model model, @PathVariable String id, HttpSession session) {
+        if(session.getAttribute("municipalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/timeSlot/update/" + id);
+            return "/inicio/login";
+        }
         model.addAttribute("timeSlot", timeSlotDao.getTimeSlot(id));
         return "timeSlot/update";
     }
@@ -63,7 +77,12 @@ public class TimeSlotController {
     }
 
     @RequestMapping(value="/delete/{id}")
-    public String processDelete(@PathVariable String id) {
+    public String processDelete(Model model, @PathVariable String id, HttpSession session) {
+        if(session.getAttribute("municipalManager") ==  null) {
+            model.addAttribute("userLogin", new UserLogin() {});
+            session.setAttribute("nextUrl", "/timeSlot/delete/" + id);
+            return "/inicio/login";
+        }
         TimeSlot timeSlot = timeSlotDao.getTimeSlot(id);
         String naturalAreaName = timeSlot.getNaturalArea();
         timeSlotDao.deleteTimeSlot(id);
